requirements.txt
fastapi
uvicorn
docxtpl
jinja2
markdown
weasyprint
PyYAML
python-multipart
pydantic[email]
sqlalchemy
passlib[bcrypt]
pyjwt
python-dateutil
